NO-ISSUE: Split agentserviceconfig files
agentserviceconfig_controller is getting quite lengthy and difficult to
navigate, pull new* functions into it's own file.
