+++ internal/controller/controllers/agentserviceconfig_new.go
@@ -1,0 +1,676 @@
+/*
+Copyright 2021.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"github.com/go-openapi/swag"
+	"github.com/hashicorp/go-version"
+	routev1 "github.com/openshift/api/route/v1"
+	aiv1beta1 "github.com/openshift/assisted-service/api/v1beta1"
+	"github.com/openshift/assisted-service/internal/common"
+	"github.com/openshift/assisted-service/internal/gencrypto"
+	"github.com/openshift/assisted-service/internal/versions"
+	"github.com/openshift/assisted-service/models"
+	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+	"github.com/sirupsen/logrus"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+func (r *AgentServiceConfigReconciler) newPVC(instance *aiv1beta1.AgentServiceConfig, name string, spec corev1.PersistentVolumeClaimSpec) (*corev1.PersistentVolumeClaim, controllerutil.MutateFn) {
+	pvc := &corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: r.Namespace,
+		},
+		Spec: spec,
+	}
+
+	requests := map[corev1.ResourceName]resource.Quantity{}
+	for key, value := range spec.Resources.Requests {
+		requests[key] = value
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, pvc, r.Scheme); err != nil {
+			return err
+		}
+		// Everything else is immutable once bound.
+		pvc.Spec.Resources.Requests = requests
+		return nil
+	}
+
+	return pvc, mutateFn
+}
+
+func (r *AgentServiceConfigReconciler) newAgentService(instance *aiv1beta1.AgentServiceConfig) (*corev1.Service, controllerutil.MutateFn) {
+	svc := &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      serviceName,
+			Namespace: r.Namespace,
+		},
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, svc, r.Scheme); err != nil {
+			return err
+		}
+		addAppLabel(serviceName, &svc.ObjectMeta)
+		if svc.ObjectMeta.Annotations == nil {
+			svc.ObjectMeta.Annotations = make(map[string]string)
+		}
+		svc.ObjectMeta.Annotations["service.beta.openshift.io/serving-cert-secret-name"] = serviceName
+		if len(svc.Spec.Ports) == 0 {
+			svc.Spec.Ports = append(svc.Spec.Ports, corev1.ServicePort{})
+		}
+		svc.Spec.Ports[0].Name = serviceName
+		svc.Spec.Ports[0].Port = servicePort
+		// since intstr.FromInt() doesn't take an int32, just use what FromInt() would return
+		svc.Spec.Ports[0].TargetPort = intstr.IntOrString{Type: intstr.Int, IntVal: servicePort}
+		svc.Spec.Ports[0].Protocol = corev1.ProtocolTCP
+		svc.Spec.Selector = map[string]string{"app": serviceName}
+		svc.Spec.Type = corev1.ServiceTypeClusterIP
+		return nil
+	}
+
+	return svc, mutateFn
+}
+
+func (r *AgentServiceConfigReconciler) newAgentRoute(instance *aiv1beta1.AgentServiceConfig) (*routev1.Route, controllerutil.MutateFn) {
+	weight := int32(100)
+	route := &routev1.Route{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      serviceName,
+			Namespace: r.Namespace,
+		},
+	}
+	routeSpec := routev1.RouteSpec{
+		To: routev1.RouteTargetReference{
+			Kind:   "Service",
+			Name:   serviceName,
+			Weight: &weight,
+		},
+		Port: &routev1.RoutePort{
+			TargetPort: intstr.FromString(serviceName),
+		},
+		WildcardPolicy: routev1.WildcardPolicyNone,
+		TLS:            &routev1.TLSConfig{Termination: routev1.TLSTerminationReencrypt},
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, route, r.Scheme); err != nil {
+			return err
+		}
+		// Only update what is specified above in routeSpec.
+		// If we update the entire route.Spec with
+		// route.Spec = routeSpec
+		// it would overwrite any existing values for route.Spec.Host
+		route.Spec.To = routeSpec.To
+		route.Spec.Port = routeSpec.Port
+		route.Spec.WildcardPolicy = routeSpec.WildcardPolicy
+		route.Spec.TLS = routeSpec.TLS
+		return nil
+	}
+
+	return route, mutateFn
+}
+
+func (r *AgentServiceConfigReconciler) newServiceMonitor(instance *aiv1beta1.AgentServiceConfig, service *corev1.Service) (*monitoringv1.ServiceMonitor, controllerutil.MutateFn) {
+	endpoints := make([]monitoringv1.Endpoint, len(service.Spec.Ports))
+	for _, port := range service.Spec.Ports {
+		endpoints = append(endpoints, monitoringv1.Endpoint{Port: port.Name})
+	}
+
+	labels := make(map[string]string)
+	for k, v := range service.ObjectMeta.Labels {
+		labels[k] = v
+	}
+
+	smSpec := monitoringv1.ServiceMonitorSpec{
+		Selector: metav1.LabelSelector{
+			MatchLabels: labels,
+		},
+		Endpoints: endpoints,
+	}
+
+	sm := &monitoringv1.ServiceMonitor{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      service.ObjectMeta.Name,
+			Namespace: r.Namespace,
+			Labels:    labels,
+		},
+		Spec: smSpec,
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, sm, r.Scheme); err != nil {
+			return err
+		}
+
+		sm.Spec = smSpec
+		sm.ObjectMeta.Labels = labels
+		return nil
+	}
+
+	return sm, mutateFn
+}
+
+func (r *AgentServiceConfigReconciler) newAgentLocalAuthSecret(instance *aiv1beta1.AgentServiceConfig) (*corev1.Secret, controllerutil.MutateFn, error) {
+	secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      agentLocalAuthSecretName,
+			Namespace: r.Namespace,
+		},
+		Type: corev1.SecretTypeOpaque,
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, secret, r.Scheme); err != nil {
+			return err
+		}
+		_, privateKeyPresent := secret.Data["ec-private-key.pem"]
+		_, publicKeyPresent := secret.Data["ec-public-key.pem"]
+		if !privateKeyPresent && !publicKeyPresent {
+			publicKey, privateKey, err := gencrypto.ECDSAKeyPairPEM()
+			if err != nil {
+				return err
+			}
+			if secret.Data == nil {
+				secret.Data = map[string][]byte{}
+			}
+			secret.Data["ec-private-key.pem"] = []byte(privateKey)
+			secret.Data["ec-public-key.pem"] = []byte(publicKey)
+		}
+		return nil
+	}
+
+	return secret, mutateFn, nil
+}
+
+func (r *AgentServiceConfigReconciler) newPostgresSecret(instance *aiv1beta1.AgentServiceConfig) (*corev1.Secret, controllerutil.MutateFn, error) {
+	pass, err := generatePassword(databasePasswordLength)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      databaseName,
+			Namespace: r.Namespace,
+		},
+		StringData: map[string]string{
+			"db.host":     "localhost",
+			"db.user":     "admin",
+			"db.password": pass,
+			"db.name":     "installer",
+			"db.port":     strconv.Itoa(int(databasePort)),
+		},
+		Type: corev1.SecretTypeOpaque,
+	}
+
+	// Only setting the owner reference to prevent clobbering the generated password.
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, secret, r.Scheme); err != nil {
+			return err
+		}
+		return nil
+	}
+
+	return secret, mutateFn, nil
+}
+
+func (r *AgentServiceConfigReconciler) newIngressCertCM(instance *aiv1beta1.AgentServiceConfig, sourceCM *corev1.ConfigMap) (*corev1.ConfigMap, controllerutil.MutateFn) {
+	cm := &corev1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      defaultIngressCertCMName,
+			Namespace: r.Namespace,
+		},
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, cm, r.Scheme); err != nil {
+			return err
+		}
+		cm.Data = make(map[string]string)
+		for k, v := range sourceCM.Data {
+			cm.Data[k] = v
+		}
+		return nil
+	}
+
+	return cm, mutateFn
+}
+
+func (r *AgentServiceConfigReconciler) newAssistedServiceDeployment(ctx context.Context, log logrus.FieldLogger, instance *aiv1beta1.AgentServiceConfig) (*appsv1.Deployment, controllerutil.MutateFn, error) {
+	var assistedConfigHash, mirrorConfigHash, userConfigHash string
+
+	// Get hash of generated assisted config
+	assistedConfigHash, err := r.getCMHash(ctx, types.NamespacedName{Name: serviceName, Namespace: r.Namespace})
+	if err != nil {
+		return nil, nil, err
+	}
+
+	envSecrets := []corev1.EnvVar{
+		// database
+		newSecretEnvVar("DB_HOST", "db.host", databaseName),
+		newSecretEnvVar("DB_NAME", "db.name", databaseName),
+		newSecretEnvVar("DB_PASS", "db.password", databaseName),
+		newSecretEnvVar("DB_PORT", "db.port", databaseName),
+		newSecretEnvVar("DB_USER", "db.user", databaseName),
+
+		// local auth secret
+		newSecretEnvVar("EC_PUBLIC_KEY_PEM", "ec-public-key.pem", agentLocalAuthSecretName),
+		newSecretEnvVar("EC_PRIVATE_KEY_PEM", "ec-private-key.pem", agentLocalAuthSecretName),
+	}
+
+	envFrom := []corev1.EnvFromSource{
+		{
+			ConfigMapRef: &corev1.ConfigMapEnvSource{
+				LocalObjectReference: corev1.LocalObjectReference{
+					Name: serviceName,
+				},
+			},
+		},
+	}
+
+	// User is responsible for:
+	// - knowing to restart assisted-service when specifying the configmap via annotation
+	// - removing the annotation when the configmap is deleted
+	userConfigName, ok := instance.ObjectMeta.GetAnnotations()[configmapAnnotation]
+	if ok {
+		log.Infof("ConfigMap %s from namespace %s being used to configure assisted-service deployment", userConfigName, r.Namespace)
+		userConfigHash, err = r.getCMHash(ctx, types.NamespacedName{Name: userConfigName, Namespace: r.Namespace})
+		if err != nil {
+			return nil, nil, err
+		}
+
+		envFrom = append(envFrom, []corev1.EnvFromSource{
+			{
+				ConfigMapRef: &corev1.ConfigMapEnvSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: userConfigName,
+					},
+				},
+			},
+		}...,
+		)
+	}
+
+	serviceContainer := corev1.Container{
+		Name:  serviceName,
+		Image: ServiceImage(),
+		Ports: []corev1.ContainerPort{
+			{
+				ContainerPort: servicePort,
+				Protocol:      corev1.ProtocolTCP,
+			},
+		},
+		EnvFrom: envFrom,
+		Env:     envSecrets,
+		VolumeMounts: []corev1.VolumeMount{
+			{Name: "bucket-filesystem", MountPath: "/data"},
+			{Name: "tls-certs", MountPath: "/etc/assisted-tls-config"},
+			{Name: "ingress-cert", MountPath: "/etc/assisted-ingress-cert"},
+		},
+		Resources: corev1.ResourceRequirements{
+			Requests: corev1.ResourceList{
+				corev1.ResourceCPU:    resource.MustParse("200m"),
+				corev1.ResourceMemory: resource.MustParse("512Mi"),
+			},
+		},
+		LivenessProbe: &corev1.Probe{
+			InitialDelaySeconds: 30,
+			Handler: corev1.Handler{
+				HTTPGet: &corev1.HTTPGetAction{
+					Path:   "/health",
+					Port:   intstr.FromInt(int(servicePort)),
+					Scheme: corev1.URISchemeHTTPS,
+				},
+			},
+		},
+		ReadinessProbe: &corev1.Probe{
+			Handler: corev1.Handler{
+				HTTPGet: &corev1.HTTPGetAction{
+					Path:   "/ready",
+					Port:   intstr.FromInt(int(servicePort)),
+					Scheme: corev1.URISchemeHTTPS,
+				},
+			},
+		},
+	}
+
+	postgresContainer := corev1.Container{
+		Name:            databaseName,
+		Image:           DatabaseImage(),
+		ImagePullPolicy: corev1.PullIfNotPresent,
+		Ports: []corev1.ContainerPort{
+			{
+				Name:          databaseName,
+				ContainerPort: databasePort,
+				Protocol:      corev1.ProtocolTCP,
+			},
+		},
+		Env: []corev1.EnvVar{
+			newSecretEnvVar("POSTGRESQL_DATABASE", "db.name", databaseName),
+			newSecretEnvVar("POSTGRESQL_USER", "db.user", databaseName),
+			newSecretEnvVar("POSTGRESQL_PASSWORD", "db.password", databaseName),
+		},
+		VolumeMounts: []corev1.VolumeMount{
+			{
+				Name:      "postgresdb",
+				MountPath: "/var/lib/pgsql/data",
+			},
+		},
+		Resources: corev1.ResourceRequirements{
+			Requests: corev1.ResourceList{
+				corev1.ResourceCPU:    resource.MustParse("100m"),
+				corev1.ResourceMemory: resource.MustParse("400Mi"),
+			},
+		},
+	}
+
+	volumes := []corev1.Volume{
+		{
+			Name: "bucket-filesystem",
+			VolumeSource: corev1.VolumeSource{
+				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
+					ClaimName: serviceName,
+				},
+			},
+		},
+		{
+			Name: "postgresdb",
+			VolumeSource: corev1.VolumeSource{
+				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
+					ClaimName: databaseName,
+				},
+			},
+		},
+		{
+			Name: "tls-certs",
+			VolumeSource: corev1.VolumeSource{
+				Secret: &corev1.SecretVolumeSource{
+					SecretName: serviceName,
+				},
+			},
+		},
+		{
+			Name: "ingress-cert",
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: defaultIngressCertCMName,
+					},
+				},
+			},
+		},
+	}
+
+	if instance.Spec.MirrorRegistryRef != nil {
+		cm := &corev1.ConfigMap{}
+		err := r.Get(ctx, types.NamespacedName{Name: instance.Spec.MirrorRegistryRef.Name, Namespace: r.Namespace}, cm)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		mirrorConfigHash, err = checksumMap(cm.Data)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		// require that the registries config key be specified before continuing
+		if _, ok := cm.Data[mirrorRegistryRefRegistryConfKey]; !ok {
+			err = fmt.Errorf("Mirror registry configmap %s missing key %s", instance.Spec.MirrorRegistryRef.Name, mirrorRegistryRefRegistryConfKey)
+			return nil, nil, err
+		}
+
+		volume := corev1.Volume{
+			Name: mirrorRegistryConfigVolume,
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: *instance.Spec.MirrorRegistryRef,
+					DefaultMode:          swag.Int32(420),
+					Items: []corev1.KeyToPath{
+						{
+							Key:  mirrorRegistryRefRegistryConfKey,
+							Path: common.MirrorRegistriesConfigFile,
+						},
+					},
+				},
+			},
+		}
+
+		serviceContainer.VolumeMounts = append(
+			serviceContainer.VolumeMounts,
+			corev1.VolumeMount{
+				Name:      mirrorRegistryConfigVolume,
+				MountPath: common.MirrorRegistriesConfigDir,
+			},
+		)
+
+		if _, ok := cm.Data[mirrorRegistryRefCertKey]; ok {
+			volume.VolumeSource.ConfigMap.Items = append(
+				volume.VolumeSource.ConfigMap.Items,
+				corev1.KeyToPath{
+					Key:  mirrorRegistryRefCertKey,
+					Path: common.MirrorRegistriesCertificateFile,
+				},
+			)
+
+			serviceContainer.VolumeMounts = append(
+				serviceContainer.VolumeMounts,
+				corev1.VolumeMount{
+					Name:      mirrorRegistryConfigVolume,
+					MountPath: common.MirrorRegistriesCertificatePath,
+					SubPath:   common.MirrorRegistriesCertificateFile,
+				},
+			)
+		}
+
+		// add our mirror registry config to volumes
+		volumes = append(volumes, volume)
+	}
+
+	deploymentLabels := map[string]string{
+		"app": serviceName,
+	}
+
+	deploymentStrategy := appsv1.DeploymentStrategy{
+		Type: appsv1.RecreateDeploymentStrategyType,
+	}
+
+	serviceAccountName := ServiceAccountName()
+
+	deployment := &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      serviceName,
+			Namespace: r.Namespace,
+		},
+		Spec: appsv1.DeploymentSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: deploymentLabels,
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: deploymentLabels,
+					Name:   serviceName,
+				},
+			},
+		},
+	}
+
+	mutateFn := func() error {
+		if err := controllerutil.SetControllerReference(instance, deployment, r.Scheme); err != nil {
+			return err
+		}
+		var replicas int32 = 1
+		deployment.Spec.Replicas = &replicas
+		deployment.Spec.Strategy = deploymentStrategy
+
+		// Handle our hashed configMap(s)
+		if deployment.Spec.Template.Annotations == nil {
+			deployment.Spec.Template.Annotations = make(map[string]string)
+		}
+		deployment.Spec.Template.Annotations[assistedConfigHashAnnotation] = assistedConfigHash
+		deployment.Spec.Template.Annotations[mirrorConfigHashAnnotation] = mirrorConfigHash
+		deployment.Spec.Template.Annotations[userConfigHashAnnotation] = userConfigHash
+
+		deployment.Spec.Template.Spec.Containers = []corev1.Container{serviceContainer, postgresContainer}
+		deployment.Spec.Template.Spec.Volumes = volumes
+		deployment.Spec.Template.Spec.ServiceAccountName = serviceAccountName
+
+		return nil
+	}
+	return deployment, mutateFn, nil
+}
+
+// getMustGatherImages returns the value of MUST_GATHER_IMAGES variable
+// to be stored in the service's ConfigMap
+//
+// 1. If mustGatherImages field is not present in the AgentServiceConfig's Spec
+//    it returns the value of MUST_GATHER_IMAGES env variable. This is also the
+//    fallback behavior in case of a processing error
+//
+// 2. If mustGatherImages field is present in the AgentServiceConfig's Spec it
+//    converts the structure to the one that can be recognize by the service
+//    and returns it as a JSON string
+//
+// 3. In case both sources are present, the Spec values overrides the env
+//    values
+func (r *AgentServiceConfigReconciler) getMustGatherImages(log logrus.FieldLogger, instance *aiv1beta1.AgentServiceConfig) string {
+	if instance.Spec.MustGatherImages == nil {
+		return MustGatherImages()
+	}
+	mustGatherVersions := make(versions.MustGatherVersions)
+	for _, specImage := range instance.Spec.MustGatherImages {
+		versionKey, err := getVersionKey(specImage.OpenshiftVersion)
+		if err != nil {
+			log.Error(err, fmt.Sprintf("Problem parsing OpenShift version %v, skipping.", specImage.OpenshiftVersion))
+			continue
+		}
+		if mustGatherVersions[versionKey] == nil {
+			mustGatherVersions[versionKey] = make(versions.MustGatherVersion)
+		}
+		mustGatherVersions[versionKey][specImage.Name] = specImage.Url
+	}
+
+	if len(mustGatherVersions) == 0 {
+		return MustGatherImages()
+	}
+	encodedVersions, err := json.Marshal(mustGatherVersions)
+	if err != nil {
+		log.Error(err, fmt.Sprintf("Problem marshaling must gather images (%v) to string, returning default %v", mustGatherVersions, MustGatherImages()))
+		return MustGatherImages()
+	}
+
+	return string(encodedVersions)
+}
+
+// getOpenshiftVersions returns the value of OPENSHIFT_VERSIONS variable
+// to be stored in the service's ConfigMap
+//
+// 1. If osImages field is not present in the AgentServiceConfig's Spec
+//    it returns the value of OPENSHIFT_VERSIONS env variable. This is
+//    also the fallback behavior in case there are no valid versions
+//    in the Spec
+//
+// 2. if osImages field is present in the AgentServiceConfig's Spec it
+//    converts the structure to the one that can be recognize by the service
+//    and returns it as a JSON string
+//
+// 3. In case both sources are present, the Spec values overrides the env
+//    values
+func (r *AgentServiceConfigReconciler) getOpenshiftVersions(log logrus.FieldLogger, instance *aiv1beta1.AgentServiceConfig) string {
+	if instance.Spec.OSImages == nil {
+		return OpenshiftVersions()
+	}
+
+	openshiftVersions := make(models.OpenshiftVersions)
+	for i, image := range instance.Spec.OSImages {
+		key, err := getVersionKey(image.OpenshiftVersion)
+		if err != nil {
+			log.Error(err, fmt.Sprintf("Problem parsing OpenShift version %v, skipping.", image.OpenshiftVersion))
+			continue
+		}
+
+		openshiftVersion := models.OpenshiftVersion{
+			DisplayName:  &key,
+			RhcosVersion: &instance.Spec.OSImages[i].Version,
+			RhcosImage:   &instance.Spec.OSImages[i].Url,
+			RhcosRootfs:  &instance.Spec.OSImages[i].RootFSUrl,
+		}
+
+		// the last entry for a particular OpenShift version takes precedence.
+		openshiftVersions[key] = openshiftVersion
+	}
+
+	if len(openshiftVersions) == 0 {
+		log.Info("No valid OS Image specified, returning default", "OpenShift Versions", OpenshiftVersions())
+		return OpenshiftVersions()
+	}
+
+	encodedVersions, err := json.Marshal(openshiftVersions)
+	if err != nil {
+		log.Error(err, fmt.Sprintf("Problem marshaling versions (%v) to string, returning default %v", openshiftVersions, OpenshiftVersions()))
+		return OpenshiftVersions()
+	}
+
+	return string(encodedVersions)
+}
+
+func (r *AgentServiceConfigReconciler) getCMHash(ctx context.Context, namespacedName types.NamespacedName) (string, error) {
+	cm := &corev1.ConfigMap{}
+	if err := r.Get(ctx, namespacedName, cm); err != nil {
+		r.Log.Error(err, "Failed to get configmap", "NamespacedName", namespacedName)
+		return "", err
+	}
+	return checksumMap(cm.Data)
+}
+
+func getVersionKey(openshiftVersion string) (string, error) {
+	v, err := version.NewVersion(openshiftVersion)
+	if err != nil {
+		return openshiftVersion, err
+	}
+
+	// put string in x.y format
+	return fmt.Sprintf("%d.%d", v.Segments()[0], v.Segments()[1]), nil
+}
+
+func newSecretEnvVar(name, key, secretName string) corev1.EnvVar {
+	return corev1.EnvVar{
+		Name: name,
+		ValueFrom: &corev1.EnvVarSource{
+			SecretKeyRef: &corev1.SecretKeySelector{
+				Key: key,
+				LocalObjectReference: corev1.LocalObjectReference{
+					Name: secretName,
+				},
+			},
+		},
+	}
+}
